Move binop codegen tests into their own module.
